+++ src/main/scala/com/rfna/checking_account/models/Statement.scala
@@ -2,4 +2,4 @@ package com.rfna.checking_account.models
 
 import java.time.LocalDate
 
-case class Statement(date: LocalDate, operations: List[Operation], balance: Double)
+case class Statement(date: LocalDate, operations: List[Operation], balance: BigDecimal)
